+++ rasa_nlu/converters.py
@@ -7,6 +7,7 @@ import io
 import json
 import logging
 
+import os
 from typing import Any
 from typing import Dict
 from typing import List
@@ -14,47 +15,62 @@ from typing import Optional
 from typing import Text
 
 from rasa_nlu import utils
-from rasa_nlu.tokenizers import Tokenizer
 from rasa_nlu.training_data import TrainingData, Message
 
 logger = logging.getLogger(__name__)
 
 # Different supported file formats and their identifier
 WIT_FILE_FORMAT = "wit"
-API_FILE_FORMAT = "api"
+DIALOGFLOW_FILE_FORMAT = "dialogflow"
 LUIS_FILE_FORMAT = "luis"
 RASA_FILE_FORMAT = "rasa_nlu"
 UNK_FILE_FORMAT = "unk"
 MARKDOWN_FILE_FORMAT = "md"
 
 
-def load_api_data(files):
+def _read_json_from_file(filename):
+    with io.open(filename, encoding="utf-8-sig") as f:
+        try:
+            return json.loads(f.read())
+        except Exception as e:
+            raise Exception("Failed to read json from '{}'. Error: "
+                            "{}".format(os.path.abspath(filename), e))
+
+
+def load_dialogflow_data(files, language):
     # type: (List[Text]) -> TrainingData
-    """Loads training data stored in the API.ai data format."""
+    """Loads training data stored in the Dialogflow data format."""
 
     training_examples = []
     entity_synonyms = {}
     for filename in files:
-        with io.open(filename, encoding="utf-8-sig") as f:
-            data = json.loads(f.read())
-        # get only intents, skip the rest. The property name is the target class
-        if "userSays" in data:
-            intent = data.get("name")
-            for s in data["userSays"]:
+        data = _read_json_from_file(filename)
+        # Language specific extensions
+        usersays_file_ext = '_usersays_{}.json'.format(language)
+        synonyms_file_ext = '_entries_{}.json'.format(language)
+        if filename.endswith(usersays_file_ext):
+            synonyms_filename = filename.replace(usersays_file_ext, '.json')
+            root_f_data = _read_json_from_file(synonyms_filename)
+            intent = root_f_data.get("name")
+
+            for s in data:
                 text = "".join([chunk["text"] for chunk in s.get("data")])
                 # add entities to each token, if available
                 entities = []
-                for e in [chunk for chunk in s.get("data") if "alias" in chunk or "meta" in chunk]:
+                for e in [chunk
+                          for chunk in s.get("data")
+                          if "alias" in chunk or "meta" in chunk]:
                     start = text.find(e["text"])
                     end = start + len(e["text"])
                     val = text[start:end]
+                    entity_type = e["alias"] if "alias" in e else e["meta"]
                     entities.append(
-                            {
-                                "entity": e["alias"] if "alias" in e else e["meta"],
-                                "value": val,
-                                "start": start,
-                                "end": end
-                            }
+                        {
+                            "entity": entity_type,
+                            "value": val,
+                            "start": start,
+                            "end": end
+                        }
                     )
                 data = {}
                 if intent:
@@ -63,9 +79,9 @@ def load_api_data(files):
                     data["entities"] = entities
                 training_examples.append(Message(text, data))
 
-        # create synonyms dictionary
-        if "name" in data and "entries" in data:
-            for entry in data["entries"]:
+        elif filename.endswith(synonyms_file_ext):
+            # create synonyms dictionary
+            for entry in data:
                 if "value" in entry and "synonyms" in entry:
                     for synonym in entry["synonyms"]:
                         entity_synonyms[synonym] = entry["value"]
@@ -79,17 +95,19 @@ def load_luis_data(filename):
     training_examples = []
     regex_features = []
 
-    with io.open(filename, encoding="utf-8-sig") as f:
-        data = json.loads(f.read())
+    data = _read_json_from_file(filename)
 
     # Simple check to ensure we support this luis data schema version
     if not data["luis_schema_version"].startswith("2"):
-        raise Exception("Invalid luis data schema version {}, should be 2.x.x. ".format(data["luis_schema_version"]) +
-                        "Make sure to use the latest luis version (e.g. by downloading your data again).")
+        raise Exception("Invalid luis data schema version {}, should be 2.x.x. "
+                        "Make sure to use the latest luis version "
+                        "(e.g. by downloading your data again)."
+                        "".format(data["luis_schema_version"]))
 
     for r in data.get("regex_features", []):
         if r.get("activated", False):
-            regex_features.append({"name": r.get("name"), "pattern": r.get("pattern")})
+            regex_features.append({"name": r.get("name"),
+                                   "pattern": r.get("pattern")})
 
     for s in data["utterances"]:
         text = s.get("text")
@@ -98,7 +116,10 @@ def load_luis_data(filename):
         for e in s.get("entities") or []:
             start, end = e["startPos"], e["endPos"] + 1
             val = text[start:end]
-            entities.append({"entity": e["entity"], "value": val, "start": start, "end": end})
+            entities.append({"entity": e["entity"],
+                             "value": val,
+                             "start": start,
+                             "end": end})
 
         data = {"entities": entities}
         if intent:
@@ -113,8 +134,7 @@ def load_wit_data(filename):
 
     training_examples = []
 
-    with io.open(filename, encoding="utf-8-sig") as f:
-        data = json.loads(f.read())
+    data = _read_json_from_file(filename)
     for s in data["data"]:
         entities = s.get("entities")
         if entities is None:
@@ -123,9 +143,13 @@ def load_wit_data(filename):
         intents = [e["value"] for e in entities if e["entity"] == 'intent']
         intent = intents[0].strip("\"") if intents else None
 
-        entities = [e for e in entities if ("start" in e and "end" in e and e["entity"] != 'intent')]
+        entities = [e
+                    for e in entities
+                    if ("start" in e and "end" in e and
+                        e["entity"] != 'intent')]
         for e in entities:
-            e["value"] = e["value"].strip("\"")    # for some reason wit adds additional quotes around entity values
+            # for some reason wit adds additional quotes around entity values
+            e["value"] = e["value"].strip("\"")
 
         data = {}
         if intent:
@@ -207,17 +231,18 @@ def rasa_nlu_data_schema():
 
 def validate_rasa_nlu_data(data):
     # type: (Dict[Text, Any]) -> None
-    """Validate rasa training data format to ensure proper training. Raises exception on failure."""
+    """Validate rasa training data format to ensure proper training.
+
+    Raises exception on failure."""
     from jsonschema import validate
     from jsonschema import ValidationError
 
     try:
         validate(data, rasa_nlu_data_schema())
     except ValidationError as e:
-        e.message += \
-            ". Failed to validate training data, make sure your data is valid. " + \
-            "For more information about the format visit " + \
-            "https://rasahq.github.io/rasa_nlu/dataformat.html"
+        e.message += (". Failed to validate training data, make sure your data "
+                      "is valid. For more information about the format visit "
+                      "https://rasahq.github.io/rasa_nlu/dataformat.html")
         raise e
 
 
@@ -225,8 +250,7 @@ def load_rasa_data(filename):
     # type: (Text) -> TrainingData
     """Loads training data stored in the rasa NLU data format."""
 
-    with io.open(filename, encoding="utf-8-sig") as f:
-        data = json.loads(f.read())
+    data = _read_json_from_file(filename)
     validate_rasa_nlu_data(data)
 
     common = data['rasa_nlu_data'].get("common_examples", list())
@@ -238,17 +262,17 @@ def load_rasa_data(filename):
     entity_synonyms = get_entity_synonyms_dict(synonyms)
 
     if intent or entity:
-        logger.warn("DEPRECATION warning: Data file contains 'intent_examples' or 'entity_examples' which will be " +
-                    "removed in the future. Consider putting all your examples into the 'common_examples' section.")
+        logger.warn("DEPRECATION warning: Data file contains 'intent_examples' "
+                    "or 'entity_examples' which will be "
+                    "removed in the future. Consider putting all your examples "
+                    "into the 'common_examples' section.")
 
     all_examples = common + intent + entity
     training_examples = []
     for e in all_examples:
-        data = {}
-        if e.get("intent"):
-            data["intent"] = e["intent"]
-        if e.get("entities") is not None:
-            data["entities"] = e["entities"]
+        data = e.copy()
+        if "text" in data:
+            del data["text"]
         training_examples.append(Message(e["text"], data))
 
     return TrainingData(training_examples, entity_synonyms, regex_features)
@@ -271,6 +295,7 @@ def guess_format(files):
 
     for filename in files:
         with io.open(filename, encoding="utf-8-sig") as f:
+            raw_data = ""
             try:
                 raw_data = f.read()
                 file_data = json.loads(raw_data)
@@ -278,8 +303,8 @@ def guess_format(files):
                     return WIT_FILE_FORMAT
                 elif "luis_schema_version" in file_data:
                     return LUIS_FILE_FORMAT
-                elif "userSays" in file_data:
-                    return API_FILE_FORMAT
+                elif "supportedLanguages" in file_data:
+                    return DIALOGFLOW_FILE_FORMAT
                 elif "rasa_nlu_data" in file_data:
                     return RASA_FILE_FORMAT
 
@@ -292,34 +317,41 @@ def guess_format(files):
 
 def resolve_data_files(resource_name):
     # type: (Text) -> List[Text]
-    """Lists all data files of the resource name (might be a file or directory)."""
+    """Lists all data files of the resource name
+
+    (might be a file or directory)."""
 
     try:
         return utils.recursively_find_files(resource_name)
     except ValueError as e:
-        raise ValueError("Invalid training data file / folder specified. {}".format(e))
+        raise ValueError("Invalid training data file / folder specified. "
+                         "{}".format(e))
 
 
-def load_data(resource_name, fformat=None):
+def load_data(resource_name, language='en', fformat=None):
     # type: (Text, Optional[Text]) -> TrainingData
-    """Loads training data from disk. If no format is provided, the format will be guessed based on the files."""
+    """Loads training data from disk.
+
+    If no format is provided, the format will be guessed based on the files."""
 
     files = resolve_data_files(resource_name)
 
     if not fformat:
         fformat = guess_format(files)
 
-    logger.info("Training data format at {} is {}".format(resource_name, fformat))
+    logger.info("Training data format at {} is {}".format(resource_name,
+                                                          fformat))
 
     if fformat == LUIS_FILE_FORMAT:
         return load_luis_data(files[0])
     elif fformat == WIT_FILE_FORMAT:
         return load_wit_data(files[0])
-    elif fformat == API_FILE_FORMAT:
-        return load_api_data(files)
+    elif fformat == DIALOGFLOW_FILE_FORMAT:
+        return load_dialogflow_data(files, language)
     elif fformat == RASA_FILE_FORMAT:
         return load_rasa_data(files[0])
     elif fformat == MARKDOWN_FILE_FORMAT:
         return load_markdown_data(files[0])
     else:
-        raise ValueError("unknown training file format : {} for file {}".format(fformat, resource_name))
+        raise ValueError("unknown training file format : {} for "
+                         "file {}".format(fformat, resource_name))