+++ rasa_nlu/data_router.py
@@ -4,6 +4,7 @@ from __future__ import print_function
 from __future__ import unicode_literals
 
 import datetime
+import glob
 import io
 import logging
 import os
@@ -78,7 +79,6 @@ class DataRouter(object):
         self.config = config
         self.responses = self._create_query_logger(config)
         self.model_dir = config['path']
-        self.token = config['token']
         self.emulator = self._create_emulator()
         self.component_builder = component_builder if component_builder else ComponentBuilder(use_cache=True)
         self.project_store = self._create_project_store()
@@ -138,9 +138,9 @@ class DataRouter(object):
         elif mode.lower() == 'luis':
             from rasa_nlu.emulators.luis import LUISEmulator
             return LUISEmulator()
-        elif mode.lower() == 'api':
-            from rasa_nlu.emulators.api import ApiEmulator
-            return ApiEmulator()
+        elif mode.lower() == 'dialogflow':
+            from rasa_nlu.emulators.dialogflow import DialogflowEmulator
+            return DialogflowEmulator()
         else:
             raise ValueError("unknown mode : {0}".format(mode))
 
@@ -152,7 +152,7 @@ class DataRouter(object):
         model = data.get("model")
 
         if project not in self.project_store:
-            projects = os.listdir(self.config['path'])
+            projects = self._list_projects(self.config['path'])
             if project not in projects:
                 raise InvalidProjectError("No project found with name '{}'.".format(project))
             else:
@@ -164,8 +164,15 @@ class DataRouter(object):
         response, used_model = self.project_store[project].parse(data['text'], data.get('time', None), model)
 
         if self.responses:
-            self.responses.info(user_input=response, project=project, model=used_model)
+            self.responses.info('', user_input=response, project=project, model=used_model)
         return self.format_response(response)
+
+    @staticmethod
+    def _list_projects(path):
+        """List the projects in the path, ignoring hidden directories."""
+        return [fn
+                for fn in glob.glob(os.path.join(path, '*'))
+                if os.path.isdir(fn)]
 
     def format_response(self, data):
         return self.emulator.normalise_response_json(data)