+++ rasa_nlu/featurizers/spacy_featurizer.py
@@ -5,12 +5,10 @@ from __future__ import absolute_import
 
 import typing
 from typing import Any
-from typing import Dict
 from typing import List
 from typing import Text
 
 from rasa_nlu.featurizers import Featurizer
-from rasa_nlu.components import Component
 from rasa_nlu.training_data import Message
 from rasa_nlu.training_data import TrainingData
 
@@ -19,6 +17,30 @@ if typing.TYPE_CHECKING:
     from spacy.language import Language
     from spacy.tokens import Doc
     import numpy as np
+
+
+def ndim(spacy_nlp):
+    """Number of features used to represent a document / sentence."""
+    # type: (Language) -> int
+    return spacy_nlp.vocab.vectors_length
+
+
+def features_for_doc(doc):
+    """Feature vector for a single document / sentence."""
+    # type: (Doc) -> np.ndarray
+    return doc.vector
+
+
+def features_for_sentences(sentences, spacy_nlp):
+    """Return a feature matrix where each row represents one sentence."""
+    # type: (List[Text], Language, int) -> np.ndarray
+    import numpy as np
+
+    X = np.zeros((len(sentences), ndim(spacy_nlp)))
+    for idx, sentence in enumerate(sentences):
+        doc = spacy_nlp(sentence)
+        X[idx, :] = features_for_doc(doc)
+    return X
 
 
 class SpacyFeaturizer(Featurizer):
@@ -32,16 +54,16 @@ class SpacyFeaturizer(Featurizer):
         # type: (TrainingData) -> None
 
         for example in training_data.intent_examples:
-            features = self.features_for_doc(example.get("spacy_doc"))
-            example.set("text_features", self._combine_with_existing_text_features(example, features))
+            self._set_spacy_features(example)
 
     def process(self, message, **kwargs):
         # type: (Message, **Any) -> None
 
-        features = self.features_for_doc(message.get("spacy_doc"))
-        message.set("text_features", self._combine_with_existing_text_features(message, features))
+        self._set_spacy_features(message)
 
-    def features_for_doc(self, doc):
-        # type: (Doc) -> np.ndarray
+    def _set_spacy_features(self, message):
+        """Adds the spacy word vectors to the messages text features."""
 
-        return doc.vector   # this will return the sentence embedding as calculated by spacy (currently averaged words)
+        fs = features_for_doc(message.get("spacy_doc"))
+        features = self._combine_with_existing_text_features(message, fs)
+        message.set("text_features", features)