+++ rasa_nlu/server.py
@@ -6,6 +6,7 @@ from __future__ import absolute_import
 import argparse
 import logging
 import os
+import six
 from functools import wraps
 
 import simplejson
@@ -31,7 +32,7 @@ def create_argparser():
                              "also be passed via a (json-formatted) config "
                              "file. NB command line args take precedence")
     parser.add_argument('-e', '--emulate',
-                        choices=['wit', 'luis', 'api'],
+                        choices=['wit', 'luis', 'dialogflow'],
                         help='which service to emulate (default: None i.e. use '
                              'simple built in format)')
     parser.add_argument('-l', '--language',
@@ -74,7 +75,10 @@ def check_cors(f):
                 request.setResponseCode(403)
                 return 'forbidden'
 
-        return f(*args, **kwargs)
+        if request.method.decode('utf-8', 'strict') == 'OPTIONS':
+            return ''  # if this is an options call we skip running `f`
+        else:
+            return f(*args, **kwargs)
 
     return decorated
 
@@ -86,9 +90,11 @@ def requires_auth(f):
     def decorated(*args, **kwargs):
         self = args[0]
         request = args[1]
-        token = str(request.args.get('token', [''])[0])
-
-        if self.data_router.token is None or token == self.data_router.token:
+        if six.PY3:
+            token = request.args.get(b'token', [b''])[0].decode("utf8")
+        else:
+            token = str(request.args.get('token', [''])[0])
+        if self.config['token'] is None or token == self.config['token']:
             return f(*args, **kwargs)
         request.setResponseCode(401)
         return 'unauthorized'
@@ -116,13 +122,13 @@ class RasaNLU(object):
     def _create_data_router(self, config, component_builder):
         return DataRouter(config, component_builder)
 
-    @app.route("/", methods=['GET'])
+    @app.route("/", methods=['GET', 'OPTIONS'])
     @check_cors
     def hello(self, request):
         """Main Rasa route to check if the server is online"""
         return "hello from Rasa NLU: " + __version__
 
-    @app.route("/parse", methods=['GET', 'POST'])
+    @app.route("/parse", methods=['GET', 'POST', 'OPTIONS'])
     @requires_auth
     @check_cors
     @inlineCallbacks
@@ -155,9 +161,10 @@ class RasaNLU(object):
                 returnValue(simplejson.dumps({"error": "{}".format(e)}))
             except Exception as e:
                 request.setResponseCode(500)
+                logger.exception(e)
                 returnValue(simplejson.dumps({"error": "{}".format(e)}))
 
-    @app.route("/version", methods=['GET'])
+    @app.route("/version", methods=['GET', 'OPTIONS'])
     @requires_auth
     @check_cors
     def version(self, request):
@@ -166,7 +173,7 @@ class RasaNLU(object):
         request.setHeader('Content-Type', 'application/json')
         return simplejson.dumps({'version': __version__})
 
-    @app.route("/config", methods=['GET'])
+    @app.route("/config", methods=['GET', 'OPTIONS'])
     @requires_auth
     @check_cors
     def rasaconfig(self, request):
@@ -175,14 +182,14 @@ class RasaNLU(object):
         request.setHeader('Content-Type', 'application/json')
         return simplejson.dumps(self.config.as_dict())
 
-    @app.route("/status", methods=['GET'])
+    @app.route("/status", methods=['GET', 'OPTIONS'])
     @requires_auth
     @check_cors
     def status(self, request):
         request.setHeader('Content-Type', 'application/json')
         return simplejson.dumps(self.data_router.get_status())
 
-    @app.route("/train", methods=['POST'])
+    @app.route("/train", methods=['POST', 'OPTIONS'])
     @requires_auth
     @check_cors
     @inlineCallbacks